removed the chiseled folder for ease of editing
