+++ gradle.properties
@@ -2,19 +2,12 @@
 org.gradle.jvmargs=-Xmx1G
 org.gradle.parallel=true
 
-# Fabric Properties
-# check these on https://fabricmc.net/develop
-minecraft_version=26.1
-fabric_loader_version=0.18.4
-loom_version=1.15-SNAPSHOT
-
-
 # Mod Properties
 mod_version=3.0.0
 maven_group=haage
 archives_base_name=locator-heads
 
-# Dependencies
-fabric_api_version=0.144.0+26.1
-cloth_config_version=26.1.154
-modmenu_version=18.0.0-alpha.8
+# Stonecutter: Active build targets (single source of truth)
+# Each version covers its hotfix range — see versions/*/gradle.properties for details.
+# 26.2 is staged in versions/26.2/ but NOT listed here until Fabric support is ready.
+stonecutter_versions=1.21.7, 1.21.9, 1.21.11, 26.1
