+++ src/main/java/haage/config/LocatorHeadsConfigManager.java
