+++ src/main/java/haage/mixin/LocatorBarRendererMixin.java
