+++ src/main/resources/assets/locator-heads/icon.png
