+++ stonecutter.gradle.kts
@@ -1,4 +1,4 @@
 plugins {
     id("dev.kikugie.stonecutter")
 }
-stonecutter active "1.21.10"
+stonecutter active "1.21.9"
