--- versions/1.21.10/gradle.properties
@@ -1,5 +1,0 @@
-minecraft_version=1.21.10
-fabric_api_version=0.138.4+1.21.10
-fabric_loader_version=0.17.3
-cloth_config_version=20.0.149
-modmenu_version=16.0.1
