+++ versions/1.21.11/gradle.properties
@@ -1,5 +1,9 @@
+# Fabric Properties — 1.21.11 standalone (major rendering + registry changes)
 minecraft_version=1.21.11
+fabric_loader_version=0.16.13
+
+# Dependencies
 fabric_api_version=0.141.3+1.21.11
-fabric_loader_version=0.18.1
 cloth_config_version=21.11.153
 modmenu_version=17.0.0
+minecraft_dependency=>=1.21.11
