+++ versions/1.21.7/gradle.properties
@@ -1,5 +1,9 @@
+# Fabric Properties — covers 1.21.7 and 1.21.8 (hotfix)
 minecraft_version=1.21.7
+fabric_loader_version=0.16.13
+
+# Dependencies
 fabric_api_version=0.129.0+1.21.7
-fabric_loader_version=0.19.2
 cloth_config_version=21.11.153
 modmenu_version=15.0.2
+minecraft_dependency=>=1.21.7 <=1.21.8
