+++ versions/1.21.9/gradle.properties
@@ -1,0 +1,9 @@
+# Fabric Properties — covers 1.21.9 and 1.21.10 (hotfix)
+minecraft_version=1.21.9
+fabric_loader_version=0.16.13
+
+# Dependencies
+fabric_api_version=0.134.1+1.21.9
+cloth_config_version=20.0.149
+modmenu_version=16.0.1
+minecraft_dependency=>=1.21.9 <=1.21.10
