+++ versions/26.1/gradle.properties
@@ -1,5 +1,9 @@
+# Fabric Properties — covers 26.1, 26.1.1, and 26.1.2 (hotfixes)
 minecraft_version=26.1
-fabric_api_version=0.144.4+26.1
 fabric_loader_version=0.18.4
+
+# Dependencies
+fabric_api_version=0.144.4+26.1
 cloth_config_version=26.1.154
 modmenu_version=18.0.0-alpha.8
+minecraft_dependency=>=26.1 <=26.1.2
