+++ versions/26.2/gradle.properties
@@ -1,5 +1,11 @@
+# Fabric Properties — 26.2 STAGING (not yet released)
+# This version is excluded from production builds via stonecutter_versions.
+# When Fabric API / loader support becomes available, add "26.2" to
+# stonecutter_versions in the root gradle.properties and update deps below.
 minecraft_version=26.2
-fabric_api_version=0.146.1+26.1.2
 fabric_loader_version=0.18.4
+
+# Dependencies (placeholder — update when 26.2 Fabric artifacts are published)
+fabric_api_version=0.146.1+26.1.2
 cloth_config_version=26.1.154
 modmenu_version=18.0.0-alpha.8
